Add ability to reuse existing OpenStack set-up
This change adds two optional flags to the context configuration for heat:
* no_setup -- to re-use an existing OpenStack deployment
* no_teardown -- to skip undeploying the context, so it is available in other testcases

In order to achieve this, resource naming had to be made more predictable.
When no_setup or no_teardown are set:
* Do not add a suffix to context name
* When searching, try with and without a suffix to find a server by name

All behaviors remain the same unless these values are set.

JIRA: YARDSTICK-886

+++ yardstick/benchmark/contexts/heat.py
@@ -332,7 +332,14 @@ class HeatContext(Context):
         if self.template_file is None:
             self._add_resources_to_template(heat_template)
 
-        self.stack = self._create_new_stack(heat_template)
+        if self._flags.no_setup:
+            # Try to get an existing stack, returns a stack or None
+            self.stack = self._retrieve_existing_stack(self.name)
+            if not self.stack:
+                self.stack = self._create_new_stack(heat_template)
+
+        else:
+            self.stack = self._create_new_stack(heat_template)
 
         # TODO: use Neutron to get segmentation-id
         self.get_neutron_info()
@@ -398,6 +405,10 @@ class HeatContext(Context):
 
     def undeploy(self):
         """undeploys stack from cloud"""
+        if self._flags.no_teardown:
+            LOG.info("Undeploying context '%s' SKIP", self.name)
+            return
+
         if self.stack:
             LOG.info("Undeploying context '%s' START", self.name)
             self.stack.delete()
@@ -445,7 +456,11 @@ class HeatContext(Context):
             server.private_ip = self.stack.outputs.get(
                 attr_name.get("private_ip_attr", object()), None)
         else:
-            server = self._server_map.get(attr_name, None)
+            try:
+                server = self._server_map[attr_name]
+            except KeyError:
+                attr_name_no_suffix = attr_name.split("-")[0]
+                server = self._server_map.get(attr_name_no_suffix, None)
             if server is None:
                 return None
